Remove CropArea from CropTool to child component
